--- Dockerfile
@@ -1,2 +1,0 @@
-FROM nginx
-COPY . /usr/share/nginx/html
