+++ sites/index/index.html
@@ -5,8 +5,6 @@
 <h2>Einleitung</h2>
 <p>Willkommen bei unseren Linux Aufgaben und Lernmaterialien! Hier können Sie Ihre Kenntnisse zu verschiedenen Themenbereichen im Linux-System testen und vertiefen.</p>
 <p>Um die bestmögliche Lernerfahrung zu erzielen, empfehlen wir Ihnen, die Befehle nicht nur auf Ihrem lokalen Computer auf einem Server auszuprobieren. Dadurch können Sie sicherstellen, dass Sie die Befehle in einer realen Umgebung testen und die Ergebnisse sehen.</p>
-<h2>Hilfestellung</h2>
-<p>Falls Sie bei einer Aufgabe wirklich nicht weiterkommen, können Sie gerne auf den Linux-Bereich zukommen.</p>
 <h2>Start</h2>
 <p>Wählen Sie einen Themenbereich aus der Sidebar, um die entsprechenden Aufgaben und Lernmaterialien anzuzeigen.</p>
 {% endblock %}
