--- templates/grundlegende-befehle.html
@@ -1,83 +1,0 @@
-{% extends 'base.html' %}
-{% block page_title %}Grundlegende Befehle{% endblock %}
-{% block first_title %}Grundlegende Befehle{% endblock %}
-{% block js_file %}<script src="{{ url_for('static', filename='grundlegende-befehle.js') }}" defer></script>{% endblock %}
-{% block main %}
-{% for task in tasks %}
-  {% include 'task.html' with task=task only %}
-{% endfor %}
-
-  <div class="task">
-    <h2>Aufgabe 1: Verzeichnis wechseln</h2>
-    <p>Lernmaterial: <a href="../tutorial/cd.html" class="lernmaterial">Tutorial Aufgabe 1</a></p>
-    <p>Frage: Wie wechseln Sie in das Verzeichnis /home/user?</p>
-    <input type="text" class="input-field" id="gb1" placeholder="Geben Sie die Lösung ein">
-    <button class="button" id="gb1-button" onclick="checkAnswer('gb1', ['cd /home/user', 'cd /home/user/'])">Lösung prüfen</button>
-    <p id="gb1-result"></p>
-    </div>
-    <div class="task">
-    <p>Frage: Wie wechseln Sie in Home-Verzeichnis des aktuellen Benutzers?</p>
-    <input type="text" class="input-field" id="gb2" placeholder="Geben Sie die Lösung ein">
-    <button class="button" id="gb2-button" onclick="checkAnswer('gb2', ['cd ~', 'cd'])">Lösung prüfen</button>
-    <p id="gb2-result"></p>
-    </div>
-    <div class="task">
-    <h2>Aufgabe 2: Dateien und Verzeichnisse auflisten</h2>
-    <p>Lernmaterial: <a href="../tutorial/ls.html" class="lernmaterial">Tutorial Aufgabe 2</a></p>
-    <p>Frage: Wie zeigen Sie die Dateien im Aktuellen Verzeichnis an? (bitte nur den Befehl ohne zusätzliche optionen)</p>
-    <input type="text" class="input-field" id="gb3" placeholder="Geben Sie die Lösung ein">
-    <button class="button" id="gb3-button" onclick="checkAnswer('gb3', ['ls', 'ls .'])">Lösung prüfen</button>
-    <p id="gb3-result"></p>
-    </div>
-    <div class="task">
-    <p>Frage: Wie zeigen Sie die Dateien im Verzeichnis /usr/local/bin an? (bitte nur den Befehl ohne zusätzliche optionen)</p>
-    <input type="text" class="input-field" id="gb4" placeholder="Geben Sie die Lösung ein">
-    <button class="button" id="gb4-button" onclick="checkAnswer('gb4', ['ls /usr/local/bin', 'ls /usr/local/bin/'])">Lösung prüfen</button>
-    <p id="gb4-result"></p>
-    </div>
-    <div class="task">
-    <p>Frage: Wie zeigen Sie die Dateien inklusive aller versteckten Dateien im aktuellen Verzeichnis an?</p>
-    <input type="text" class="input-field" id="gb5" placeholder="Geben Sie die Lösung ein">
-    <button class="button" id="gb5-button" onclick="checkAnswer('gb5', ['ls -a', 'ls . -a'])">Lösung prüfen</button>
-    <p id="gb5-result"></p>
-    </div>
-    <div class="task">
-    <h2>Aufgabe 3: Aktuelles Verzeichnis anzeigen</h2>
-    <p>Lernmaterial: <a href="../tutorial/pwd.html" class="lernmaterial">Tutorial Aufgabe 3</a></p>
-    <p>Frage: Wie zeigen Sie das aktuelle Verzeichnis an, in dem Sie sich befinden?</p>
-    <input type="text" class="input-field" id="gb6" placeholder="Geben Sie die Lösung ein">
-    <button class="button" id="gb6-button" onclick="checkAnswer('gb6', ['pwd'])">Lösung prüfen</button>
-    <p id="gb6-result"></p>
-    </div>
-    <div class="task">
-    <h2>Aufgabe 4: Verzeichnisse erstellen und entfernen</h2>
-    <p>Lernmaterial: <a href="../tutorial/mkdir-rmdir.html" class="lernmaterial">Tutorial Aufgabe 4</a></p>
-    <p>Frage: Wie erstellen Sie ein neues Verzeichnis namens "MeineDokumente" im aktuellen Verzeichnis?</p>
-    <input type="text" class="input-field" id="gb7" placeholder="Geben Sie die Lösung ein">
-    <button class="button" id="gb7-button" onclick="checkAnswer('gb7', ['mkdir MeineDokumente', 'mkdir meinedokumente'])">Lösung prüfen</button>
-    <p id="gb7-result"></p>
-    </div>
-    <div class="task">
-    <p>Frage: Wie erstellen Sie ein neues Verzeichnis namens "Arbeit" im aktuellen Verzeichnis und gleichzeitig ein Unterverzeichnis namens "Projekte" innerhalb des "Arbeit"-Verzeichnisses? </p>
-    <input type="text" class="input-field" id="gb8" placeholder="Geben Sie die Lösung ein">
-    <button class="button" id="gb8-button" onclick="checkAnswer('gb8', ['mkdir -p Arbeit/Projekte', 'mkdir -p arbeit/projekte'])">Lösung prüfen</button>
-    <p id="gb8-result"></p>
-    </div>
-    <div class="task">
-    <p>Frage: Wie entfernen Sie ein leeres Verzeichnis namens "AlteDokumente" im aktuellen Verzeichnis?  </p>
-    <input type="text" class="input-field" id="gb9" placeholder="Geben Sie die Lösung ein">
-    <button class="button" id="gb9-button" onclick="checkAnswer('gb9', ['rmdir AlteDokumente', 'rmdir altedokumente'])">Lösung prüfen</button>
-    <p id="gb9-result"></p>
-    </div>
-    <div class="task">
-    <p>Frage: Wie entfernen Sie ein leeres Verzeichnis namens "AlteDokumente" und seine übergeordneten Verzeichnisse vorausgesetzt, dass alle Verzeichnisse leer sind?  </p>
-    <input type="text" class="input-field" id="gb10" placeholder="Geben Sie die Lösung ein">
-    <button class="button" id="gb10-button" onclick="checkAnswer('gb10', ['rmdir -p AlteDokumente', 'rmdir -p altedokumente'])">Lösung prüfen</button>
-    <p id="gb10-result"></p>
-    </div>
-
-<!-- Weitere Aufgaben hier hinzufügen -->
-<div>
-<button class="button" onclick="resetSolutions()">Alle Lösungen zurücksetzen</button>
-</div>
-{% endblock %}
