+++ templates/task.html
@@ -1,10 +1,15 @@
+{% if task.title %}
+  <br>
+{% endif %}
 <div class="task">
-    <h2>Aufgabe {{ task.number }}: {{ task.title }}</h2>
+    {% if task.title %}
+      <h2>Aufgabe {{ task.number }}: {{ task.title }}</h2>
+    {% endif %}
     {% if task.tutorial %}
       <p>Lernmaterial: <a href="{{ task.tutorial }}" class="lernmaterial">Tutorial Aufgabe {{ task.number }}</a></p>
     {% endif %}
     <p>Frage: {{ task.question }}</p>
     <input type="text" class="input-field" id="{{ task.input_id }}" placeholder="Geben Sie die Lösung ein">
-    <button class="button" id="{{ task.button_id }}" onclick="checkAnswer('{{ task.input_id }}', {{ task.answers|tojson }})">Lösung prüfen</button>
+    <button class="button" id="{{ task.button_id }}" onclick='checkAnswer("{{ task.input_id }}", {{ task.answers|tojson }})'>Lösung prüfen</button>
     <p id="{{ task.result_id }}"></p>
 </div>
